connection: handle sending responses in the comms
Move the responsibility of sending back responses to the
communications loop. Have the commands return the following:

int code[, str message[, bool disconnect]]

This makes it less likely to leave a conversation hanging in the
middle due to a buggy command. If a command returns, something will
be sent back to the client.